refactor(fetch): move fetch into dir mod
