+++ crates/jstz_proto/src/runtime/v2/fetch/tests/resources/fetch_cleans_headers/run.js
