+++ crates/jstz_proto/src/runtime/v2/fetch/tests/resources/transfer_fails_when_error_thrown/run.js
