+++ crates/jstz_proto/src/runtime/v2/fetch/tests/resources/transfer_succeeds/remote.js
